test: use unique names for ProducerSystemTest
